Move cluster to separate module, remove inner pool
