+++ src/lib.rs
@@ -5,3 +5,6 @@ mod connection;
 mod errors;
 mod factory;
 mod pool;
+
+#[cfg(feature = "cluster")]
+mod cluster_pool;
